requirements.txt
aniso8601==7.0.0
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
django-health-check==3.18.3
graphene==2.1.9
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2-binary==2.9.10
PyJWT==1.7.1
python-decouple==3.8
pytz==2025.2
Rx==1.6.3
setuptools==80.1.0
singledispatch==4.1.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
Unidecode==1.4.0
